+++ __tests__/setup/conditions/suggestedActionsShown.js
@@ -1,5 +1,5 @@
 import { By, until } from 'selenium-webdriver';
 
-export default function suggestedActionsShowed() {
+export default function suggestedActionsShown() {
   return until.elementLocated(By.css('[role="form"] ul'));
 }